requirements.txt
pip
Flask
cassandra-driver    --install-option="--no-cython"
Werkzeug
